Split ffi_internal.rs into per-category modules (#665)
Replace the monolithic ffi_internal.rs with a ffi_internal/ directory
containing focused sub-modules mirroring the existing ffi/ structure:
- recording.rs — discovery data recording
- analysis.rs — parallel analysis, focus neuron ranking, calibration
- gpu.rs — GPU availability probe, library version
- utilities.rs — parquet merge, record reading, visualisation export

All public API signatures unchanged. All tests moved to mod.rs.
